Move ItemGroup and SchedulerItem classes
