+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/scheduling/Strand.java
@@ -57,7 +57,6 @@ import static io.ballerina.runtime.internal.scheduling.State.YIELD;
  *
  * @since 0.955.0
  */
-
 public class Strand {
 
     private static final AtomicInteger nextStrandId = new AtomicInteger(0);
@@ -405,12 +404,8 @@ public class Strand {
         return id;
     }
 
-    public int getStrandGroupId() {
-        return strandGroup.getId();
-    }
-
-    public boolean isStrandGroupScheduled() {
-        return strandGroup.isScheduled();
+    public ItemGroup getStrandGroup() {
+        return strandGroup;
     }
 
     /**
